Improve changes as per PR18
