+++ converter/mpi_cog_convert.sh
@@ -25,7 +25,7 @@ while [[ "$#" -gt 0 ]]; do
         --config-path )         shift
                                 YAMLFILE="$1"
                                 ;;
-        --streamer-path )       shift
+        --cog-converter-path )  shift
                                 COGS="$1"
                                 ;;
         * )
@@ -74,14 +74,14 @@ cd "$OUTDIR" || exit 1
 j=1
 f_j=$(qsub -V -P "$PROJECT" -q "$QUEUE" \
       -l walltime=1:00:00,mem=$MEM,jobfs=$JOBFS,ncpus=$NCPUS,wd \
-      -- mpirun python3 "$COGS" mpi-convert-cog -c "$YAMLFILE" --output-dir "$OUTDIR" \
+      -- mpirun --tag-output python3 "$COGS" mpi-convert-cog -c "$YAMLFILE" --output-dir "$OUTDIR" \
       --product-name "$PRODUCT" "$FILEL$j")
 
 j=2
 while [ -s  "$FILEL$j" ]; do
     n_j=$(qsub -V -W depend=afterany:"$f_j" -P "$PROJECT" -q "$QUEUE" \
           -l walltime=1:00:00,mem=$MEM,jobfs=$JOBFS,ncpus=$NCPUS,wd \
-          -- mpirun python3 "$COGS" mpi-convert-cog -c "$YAMLFILE" --output-dir "$OUTDIR" \
+          -- mpirun --tag-output python3 "$COGS" mpi-convert-cog -c "$YAMLFILE" --output-dir "$OUTDIR" \
           --product-name "$PRODUCT" "$FILEL$j")
     f_j=$n_j
     j=$((j+1))
